refactor: remove polling, add DownloadsProvider and reorganise
