+++ src/main/ipc/browser/downloads.ts
@@ -2,6 +2,11 @@ import fs from "node:fs";
 import { ipcMain, shell } from "electron";
 import { downloadsController } from "@/controllers/downloads-controller";
 import { deleteDownloadRecord, getDownloadRecord, listDownloads } from "@/saving/downloads";
+import { sendMessageToListeners } from "@/ipc/listeners-manager";
+
+export function fireDownloadsChanged() {
+  sendMessageToListeners("downloads:on-changed");
+}
 
 ipcMain.handle("downloads:list", () => {
   return downloadsController.listDownloads();
@@ -38,16 +43,21 @@ ipcMain.handle("downloads:open-file", (_event, downloadId: string) => {
 });
 
 ipcMain.handle("downloads:remove-record", (_event, downloadId: string) => {
-  return deleteDownloadRecord(downloadId);
+  const ok = deleteDownloadRecord(downloadId);
+  if (ok) fireDownloadsChanged();
+  return ok;
 });
 
 ipcMain.handle("downloads:clear-completed", () => {
   const downloads = listDownloads();
+  let changed = false;
   for (const dl of downloads) {
     if (dl.state === "completed" || dl.state === "cancelled") {
       deleteDownloadRecord(dl.id);
+      changed = true;
     }
   }
+  if (changed) fireDownloadsChanged();
 });
 
 ipcMain.handle("downloads:check-files-exist", (_event, downloadIds: string[]) => {
