+++ src/renderer/src/routes/downloads/config.tsx
@@ -1,28 +1,14 @@
 import { ThemeProvider } from "@/components/main/theme";
+import { DownloadsProvider } from "@/components/downloads/manager/provider";
 import { RouteConfigType } from "@/types/routes";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { ReactNode, useState } from "react";
-
-function DownloadsQueryProvider({ children }: { children: ReactNode }) {
-  const [queryClient] = useState(
-    () =>
-      new QueryClient({
-        defaultOptions: {
-          queries: {
-            staleTime: 5_000
-          }
-        }
-      })
-  );
-  return <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>;
-}
+import { ReactNode } from "react";
 
 export const RouteConfig: RouteConfigType = {
   Providers: ({ children }: { children: ReactNode }) => {
     return (
-      <DownloadsQueryProvider>
-        <ThemeProvider forceTheme="dark">{children}</ThemeProvider>
-      </DownloadsQueryProvider>
+      <ThemeProvider forceTheme="dark">
+        <DownloadsProvider>{children}</DownloadsProvider>
+      </ThemeProvider>
     );
   }
 };
